CashFlow more flexible
CashFlow was defined in such a way that all elements of `currency`,
`date` and `amount` fields (which are all Vector types) had to be the
same Julia’s invariant type parameter system.

This restriction is no longer applicable. If a CashFlow instance has
flows all of the same currency, then the Vector containing them which
will be of type Vector{CCY} is “promoted” to Vector{Currency}. Same for
other fields.